+++ Core.cs
@@ -7,24 +7,15 @@ using Bloodcraft.Systems.Expertise;
 using Bloodcraft.Systems.Legacies;
 using Bloodcraft.Systems.Legacy;
 using Bloodcraft.Systems.Leveling;
-using Il2CppInterop.Runtime;
-using Il2CppInterop.Runtime.Injection;
-using Il2CppInterop.Runtime.InteropTypes;
-using Il2CppSystem.ComponentModel;
-using Il2CppSystem.ComponentModel.Design;
 using ProjectM;
 using ProjectM.Network;
 using ProjectM.Physics;
 using ProjectM.Scripting;
 using ProjectM.Shared.Systems;
-using ProjectM.Terrain;
 using Stunlock.Core;
 using System.Collections;
 using System.Text.Json;
-using Unity.Burst;
 using Unity.Entities;
-using Unity.Jobs;
-using Unity.Services.Core.Internal;
 using UnityEngine;
 using static Bloodcraft.Core.DataStructures;
 
@@ -88,7 +79,6 @@ internal static class Core
 
         hasInitialized = true;
     }
-    
     static World GetWorld(string name)
     {
         foreach (var world in World.s_AllWorlds)