+++ Patches/InitializationPatch.cs
@@ -1,16 +1,18 @@
 using HarmonyLib;
 using ProjectM;
+using ProjectM.Gameplay.WarEvents;
 
 namespace Bloodcraft.Patches;
 
 [HarmonyPatch]
-[HarmonyPriority(1001)]
-internal static class InitializationPatches
+internal static class InitializationPatch
 {
+    /*
     [HarmonyPatch(typeof(SpawnTeamSystem_OnPersistenceLoad), nameof(SpawnTeamSystem_OnPersistenceLoad.OnUpdate))]
     [HarmonyPostfix]
-    static void SpawnTeamSystemOnUpdatePostfix()
+    static void OnUpdatePostfix()
     {
+        Core.Log.LogInfo($"SpawnTeamSystem_OnPersistenceLoad.OnUpdate() Postfix...");
         try
         {
             Core.Initialize();
@@ -19,6 +21,18 @@ internal static class InitializationPatches
         catch (Exception e)
         {
             Core.Log.LogWarning($"{MyPluginInfo.PLUGIN_NAME}[{MyPluginInfo.PLUGIN_VERSION}] failed to initialize with {e} on try-catch (other mod initializations in SpawnTeamSystem_OnPersistenceLoad.OnUpdate() shouldn't be affected by this)");
+        }
+    }
+    */
+
+    [HarmonyPatch(typeof(WarEventRegistrySystem), nameof(WarEventRegistrySystem.RegisterWarEventEntities))]
+    [HarmonyPostfix]
+    static void Postfix()
+    {
+        if (!Core.hasInitialized) Core.Initialize();
+        if (Core.hasInitialized)
+        {
+            Core.Log.LogInfo($"|{MyPluginInfo.PLUGIN_NAME}[{MyPluginInfo.PLUGIN_VERSION}] initialized|");
         }
     }
 }